+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.0.2] - 2022-08-27
 ### Customer service
 - first test setup
 - fix Import of pydantic BaseSetting during tests